requirements.txt
numpy
pandas
matplotlib
scikit-learn
pygame
colorama
torch
pynput
pygame
pyserial
torchvision
depthai
opencv-python
pillow
roboflow
ultralytics